requirements.txt
google-adk[database]
yfinance
litellm
google-generativeai
python-dotenv